refactor(Map): Extract the user position pin to a special hook
